sync: Review files elaboration
The pkgs-checker library return empty string
if a specific metadata file is empty, this avoid
injection of all files of the package.

In order to set correctly all metadata files
inside PMS database we check if the file exists
when the content is empty.

This fix the correct uninstall of the synced
packages with DEBUGBUILD, PROPERTIES and other
emtpy files.