Permit to generate specs for multiple versions
In order to support a way to generate specs of
packages with a single SLOT but multiple versions,
for example Nodejs package we add new options
for artefact struct.

The option `custom_path` permits to define
the custom path used for the anise specs of a
conditioned package.

The option `conditions` permits to define one
or more string following the Gentoo/Funtoo specs
in order to filter a specific package version.

So, in order to use this feature you need to
define a new artefact block where append to the
package string a specific string (after the comma)
used to identify the custom package from the default
package that will match with the last version available.

For example:

```yaml

- tree: packages/atoms-extra/lang
  custom_path: net-libs/nodejs/20
  conditions:
    - '<net-libs/nodejs-22'
    - '>=net-libs/nodejs-20'
  uses:
    enabled:
      - icu
      - inspector
      - npm
      - snapshot
      - ssl
      - system-icu
      - system-ssl
    disabled:
      - debug
      - doc
      - lto
      - pax-kernel
      - systemtap
      - test
  packages:
    - net-libs/nodejs,nodejs20

```

In the example, the label used is `nodejs20`.

Closes: macaroni-os/anise-portage-converter#8